# api/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Private REST API to for user interface application",
        "title": "storePredictor private REST API",
        "contact": {
            "name": "Ales Jandera",
            "url": "http://storepredictor.com",
            "email": "ales@storepredictor.com"
        },
        "license": {
            "name": "Commercial"
        },
        "version": "1.0"
    },
    "host": "localhost:9996",
    "basePath": "/",
    "paths": {
...
        "/account/order/{accountId}/{storeId}": {
            "get": {
                "description": "Endpoint to get account orders",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Endpoint to get account orders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            },
            "post": {
                "description": "Endpoint to store Order data from account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Endpoint to store Order data from account",
                "parameters": [
                    {
                        "description": "Order object to store in database",
                        "name": "o.Order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            }
        },
        "/accounts": {
            "get": {
                "description": "Endpoint to send all accounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Endpoint to send all accounts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/structs.Account"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Endpoint to update account data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Endpoint to store account data",
                "parameters": [
                    {
                        "description": "Account object to update in database",
                        "name": "a.Account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.Account"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            }
        },
...
        "/accounts/{lang}": {
            "post": {
                "description": "Endpoint to StoreWeights data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StoreWeights"
                ],
                "summary": "Endpoint to store StoreWeights data",
                "parameters": [
                    {
                        "description": "StoreWeights object to store in database",
                        "name": "sw.StoreWeights",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.StoreWeights"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            }
        },
...
        "/restore/{token}": {
            "put": {
                "description": "Endpoint to update password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Password"
                ],
                "summary": "Endpoint to update password",
                "parameters": [
                    {
                        "description": "Pw object to update in database",
                        "name": "structs.Pw",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.Pw"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Order": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "companyNumber": {
                    "type": "string"
                },
                "countryCode": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "paidTo": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "planRefer": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "vatNumber": {
                    "type": "string"
                },
                "zip": {
                    "type": "string"
                }
            }
        },
        "structs.Account": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "companyNumber": {
                    "type": "string"
                },
                "countryCode": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "newsletter": {
                    "type": "boolean"
                },
                "newsletterConfirmation": {
                    "type": "string"
                },
                "paidTo": {
                    "type": "string"
                },
                "parent": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "planRefer": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "vatNumber": {
                    "type": "string"
                },
                "zip": {
                    "type": "string"
                }
            }
        },
...
        "structs.Pw": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
        "structs.StoreWeights": {
            "type": "object",
            "properties": {
                "a": {
                    "type": "number"
                },
                "b": {
                    "type": "number"
                },
                "beta": {
                    "type": "number"
                },
                "c": {
                    "type": "number"
                },
                "d": {
                    "type": "number"
                },
                "delta": {
                    "type": "number"
                },
                "e": {
                    "type": "number"
                },
                "gama": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "probabilityWeights": {
                    "type": "string"
                },
                "storeRefer": {
                    "type": "string"
                }
            }
        },
...
        "utils.Res": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}