# api/src/structs/order/order.go
package model

type Order struct {
	Name          string
	Email         string
	Street        string
	City          string
	Zip           string
	CountryCode   string
	CompanyNumber string
	VatNumber     string
	PaidTo        string
	PlanRefer     string
	Password      string
}